backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
boto3==1.34.34
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
email-validator==2.1.0.post1
python-multipart==0.0.9
requests==2.32.3
requests-aws4auth==1.3.1
python-dotenv==1.0.1
PyYAML==6.0.2
